requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
flask>=3.0.0
waitress>=3.0.0
werkzeug>=3.0.0
requests>=2.31.0

# ─── Computer Vision ──────────────────────────────────────────────────────────
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0

# ─── Person Detection (YOLOv8) ────────────────────────────────────────────────
ultralytics>=8.0.0

# ─── Person Tracking (DeepSORT) ───────────────────────────────────────────────
deep-sort-realtime>=1.3.2
filterpy>=1.4.5
scikit-learn>=1.3.0

# ─── Face & Emotion Analysis (DeepFace) ──────────────────────────────────────
deepface>=0.0.79
retina-face>=0.0.17
tf-keras>=2.15.0

# ─── Action Recognition (SlowFast via PyTorchVideo) ──────────────────────────
# Install PyTorch separately: see Dockerfile / install_windows.ps1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorchvideo>=0.1.5

# ─── Database ─────────────────────────────────────────────────────────────────
# Using Python's built-in sqlite3 — no external package needed
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
